#34 Refactor add new market
