+++ app/member/forms.py
@@ -3,9 +3,7 @@ from django.contrib.auth import authenticate
 from django.contrib.auth.forms import UserCreationForm
 
 import member
-from market.models import Market
 from member.models import Member
-from tag.models import Tag
 
 
 class MemberForm(UserCreationForm):
@@ -46,24 +44,3 @@ class LoginForm(forms.Form):
                 "아이디 혹은 비밀번호가 일치하지 않습니다."
             )
         return self.cleaned_data
-
-
-class MarketForm(forms.ModelForm):
-    posts = forms.CharField(label='게시글', widget=forms.Textarea, required=False)
-    tags = forms.MultipleChoiceField(
-        label='태그',
-        choices=[(ele.pk, ele.tag) for ele in Tag.objects.all().order_by('id')],
-        widget=forms.CheckboxSelectMultiple(),
-        required=False,
-    )
-
-    def clean(self):
-        studio_name = self.cleaned_data.get('studio_name')
-        if Market.objects.filter(studio_name__exact=studio_name):
-            self.add_error('studio_name', "중복된 스튜디오 이름입니다.")
-        return self.cleaned_data
-
-    class Meta:
-        model = Market
-        fields = ('studio_name', 'location', 'posts', 'costs', 'photo',
-                  'kakao_id',)
